% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_random_forest.R
\name{fit_random_forest_clustered}
\alias{fit_random_forest_clustered}
\title{Train the clustered data with random forest-based algorithms}
\usage{
fit_random_forest_clustered(train_data, test_data, method)
}
\arguments{
\item{train_data}{Training data in the 2d plane.}

\item{test_data}{Testing data.}

\item{method}{Five algorithms are included. "rf": random forest, "wrf": weighted random forest, "brf": balanced random forest, "brf1": balanced random forest with refined weights using method 1, "brf2": balanced random forest with refined weights using method 2.}
}
\value{
Return a list including `model`, the model after trianing, and `measure`, the six measures on the testing data.
}
\description{
Train the clustered data with random forest-based algorithms
}
