% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_clustered_data}
\alias{generate_clustered_data}
\title{Generate data with two groups, for which the major group comes from a Gaussian distribution and the minor group comes from a mixture of two Gaussian distributions.}
\usage{
generate_clustered_data(parameters, n_train_zero, n_train_one, n_train_two)
}
\arguments{
\item{parameters}{A list containing the mean of the major distribution, the covariance of the major distribution, the mean of the minor distribution and the covariance of the minor distribution}

\item{n_train_zero}{The sample size of the major group.}

\item{n_train_one}{The sample size of the minor group from the first distribution of the mixture of Gaussian.}

\item{n_train_two}{The sample size of the minor group from the second distribution of the mixture of Gaussian.}
}
\value{
This function returns a list including `train_data` the training data and `test_data` the testing data
}
\description{
Generate data with two groups, for which the major group comes from a Gaussian distribution and the minor group comes from a mixture of two Gaussian distributions.
}
