% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate data with two groups, each of which comes from a Gaussian distribution respectively.}
\usage{
generate_data(parameters, n_train_zero, n_train_one)
}
\arguments{
\item{parameters}{A list containing the mean of the major distribution, the covariance of the major distribution, the mean of the minor distribution and the covariance of the minor distribution}

\item{n_train_zero}{The sample size of the major group.}

\item{n_train_one}{The sample size of the minor group.}
}
\value{
This function returns a list including `train_data` the training data and `test_data` the testing data
}
\description{
Generate data with two groups, each of which comes from a Gaussian distribution respectively.
}
