% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_covariance.R
\name{get_mean_covariance}
\alias{get_mean_covariance}
\title{Get prespecified mean and covariances for the Gaussian distribution from which the data come.}
\usage{
get_mean_covariance(choice, mean_dist)
}
\arguments{
\item{choice}{Taking 1, 2, 3, or 4, this function offers four choices to allow different settings of data distribution.}

\item{mean_dist}{This argument specifies the distance between the mean of two Gaussian distribution.}
}
\value{
This function returns a list including `mean_zero`, `mean_one`, `cov_zero`, `cov_one`
}
\description{
Get prespecified mean and covariances for the Gaussian distribution from which the data come.
}
