% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_random_forest.R
\name{get_print_results}
\alias{get_print_results}
\title{Compute six measures after training}
\usage{
get_print_results(rf_model)
}
\arguments{
\item{rf_model}{The trained model.}
}
\value{
Return six measures: the classiciation accuracy on the major group, the classiciation accuracy on the minor group, precision, F-measure, G-mean, and weighted accuracy.
}
\description{
Compute six measures after training
}
