% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_random_forest.R
\name{modify_rf_with_prior}
\alias{modify_rf_with_prior}
\title{Calibrate the random forest model with prior information and test it on the testing data.}
\usage{
modify_rf_with_prior(rf_model, class_zero_ratio, test_data)
}
\arguments{
\item{rf_model}{The model after training with the ordinary random forest.}

\item{class_zero_ratio}{The value of $P(Y=0)$, which represents the ratio of data with label 0 among all the data.}

\item{test_data}{Testing data.}
}
\value{
This function returns six measures of the calibrated model with prior information on the testing data.
}
\description{
Calibrate the random forest model with prior information and test it on the testing data.
}
